refactor start command folder
