+++ packages/xdl/src/Env.ts
@@ -1,5 +1,8 @@
+import { ExpoConfig } from '@expo/config';
 import getenv from 'getenv';
 import os from 'os';
+
+import * as Versions from './Versions';
 
 export function home(): string {
   return os.homedir();
@@ -15,4 +18,13 @@ export function isLocal(): boolean {
 
 export function maySkipManifestValidation(): boolean {
   return !!getenv.string('EXPO_SKIP_MANIFEST_VALIDATION_TOKEN');
+}
+
+/**
+ * Returns true if we should use Metro using its JS APIs via @expo/dev-server (the modern and fast
+ * way), false if we should fall back to spawning it as a subprocess (supported for backwards
+ * compatibility with SDK39 and older).
+ */
+export function shouldUseDevServer(exp: ExpoConfig) {
+  return Versions.gteSdkVersion(exp, '40.0.0') || getenv.boolish('EXPO_USE_DEV_SERVER', false);
 }
