+++ packages/xdl/src/start/__tests__/fixtures/icon.png
